% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mprofile.R
\name{mprofile}
\alias{mprofile}
\title{Count motifs in a network}
\usage{
mprofile(M, normalise)
}
\arguments{
\item{M}{A numeric matrix representing interactions between two groups of species. Each row corresponds to a species in one level
and each column corresponds to a species in the other level. Elements of M are positive numbers if species do interact, and 0 
otherwise. Formally, M is an incidence matrix. When species i and j interact, m_ij > 0; if they do not interact, m_ij = 0.
If interactions are weighted (matrix elements are greater than 1), the function will automatically convert the matrix to a binary
matrix.}

\item{normalise}{Logical; should motif frequencies be normalised to control for network size?}
}
\description{
This function counts the number of times each of the 17 motifs up to five nodes occurs in a network
}
\examples{
set.seed(123)
row <- 100
col <- 100
m <- matrix(sample(0:1, row*col, replace=TRUE), row, col)
mprofile(M = m, normalise = TRUE)
}

