% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roles.R
\name{roles}
\alias{roles}
\title{Compute species' structural roles}
\usage{
roles(M, level = "all", normalisation = "none")
}
\arguments{
\item{M}{A numeric matrix representing interactions between two groups of species. Each row corresponds to a species in one level
and each column corresponds to a species in the other level. Elements of M are positive numbers if species do interact, and 0 
otherwise. Formally, M is an incidence matrix. When species i and j interact, m_ij > 0; if they do not interact, m_ij = 0.
If interactions are weighted (non-zero matrix elements can take values greater than 1), the function will automatically convert the matrix to a binary
matrix.}

\item{level}{Which species level should roles be calculated for: \code{rows}, \code{columns} or \code{all}?  Defaults to \code{all}.}

\item{normalisation}{Which normalisation should be used: \code{none}, \code{across} or \code{within}?  Defaults to \code{none}.}
}
\value{
It returns a data frame with A + P rows and 46 columns, corresponding to the 46 unique positions (Figure 1). In addition to the input network, roles has two other arguments. The ‘level’ argument allows the user to specify if they want to calculate positions for ‘row’ species only, ‘column’ species only, or all species.
}
\description{
This function counts the number of times each species in a network occurs in each of the 46 positions found within the 17 motifs up to five nodes
}
\details{
The \code{level} arguent controls which species' roles are calculated for. \code{rows} returns position counts for all species in rows, \code{columns}
returns counts for all species in columns, and \code{all} return counts for all spcies in the network.

Species with more interactions will tend to appear in more positions. Normalisation helps control for this.
\code{none} performs no normalisation and will return the raw position counts.
\code{across} divides position counts for each species by the total number of times that species appears in any position.
\code{within} divides position counts for each species by the total number of times that species appears in any position within the same motif size class.
Which normalisation is most appropriate will depend on the question being asked.
}
\examples{
set.seed(123)
row <- 100
col <- 100
m <- matrix(sample(0:1, row*col, replace=TRUE), row, col)
roles(M = m, level = "all", normalisation = "none")
}

